Don't run tearDown when not supported

+++ backend/configuration/src/integrationTest/java/com/elvaco/mvp/web/GatewayControllerTest.java
@@ -55,6 +55,9 @@ public class GatewayControllerTest extends IntegrationTest {
 
   @After
   public void tearDown() {
+    if (!isPostgresDialect()) {
+      return;
+    }
     physicalMeterJpaRepository.deleteAll();
     gatewayStatusLogJpaRepository.deleteAll();
     logicalMeterJpaRepository.deleteAll();
